Add Font Awesome libraries for Elementor theme
Added the necessary Font Awesome libraries to the Elementor theme for the Wordpress site. These libraries include "conflict-detection.min.js" and "brands.js" which are essential for loading Font Awesome icons properly on the website. This should allow a wider variety of icons to be used throughout the site's design.