Update search bar appearance and logic
Revised the search bar's appearance and logic by adding styles for a minimal skin search form. Further modifications were made on the utilities SCSS file to use a mixin for screen reader text. Fixed the logic in the dynamic header to display the appropriate search form depending on the user's settings.